refactor: 패키지 경로 수정
- domain과 global
- 공통 사용 또는 설정 관련된 것은 global 패키지로 이동
