--- src/main/java/com/example/jwt/auth/DelegatedAccessDeniedHandler.java
@@ -1,27 +1,0 @@
-package com.example.jwt.auth;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-
-import java.io.IOException;
-
-@Component
-public class DelegatedAccessDeniedHandler implements AccessDeniedHandler {
-
-    private final HandlerExceptionResolver resolver;
-
-    public DelegatedAccessDeniedHandler(@Qualifier("handlerExceptionResolver") HandlerExceptionResolver resolver) {
-        this.resolver = resolver;
-    }
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        resolver.resolveException(request, response, null, accessDeniedException);
-    }
-}
