+++ src/main/java/com/example/jwt/domain/user/dto/UserDetailResDto.java
@@ -1,0 +1,13 @@
+package com.example.jwt.domain.user.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class UserDetailResDto {
+
+    private final Long id;
+    private final String email;
+    private final String nickname;
+}
