+++ src/main/java/com/example/jwt/domain/user/entity/User.java
@@ -1,0 +1,42 @@
+package com.example.jwt.domain.user.entity;
+
+import com.example.jwt.global.common.entity.BaseEntity;
+import com.example.jwt.global.common.constant.UserRole;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.SQLDelete;
+
+@Entity
+@Table(name = "user")
+@Getter
+@NoArgsConstructor
+@SQLDelete(sql = "UPDATE user SET is_deleted = true WHERE id = ?")
+public class User extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length = 50, nullable = false, unique = true)
+    private String email;
+
+    @Column(length = 100, nullable = false)
+    private String password;
+
+    @Column(length = 20, nullable = false)
+    private String nickname;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private UserRole role;
+
+    private boolean isDeleted = false;
+
+    public User(String email, String password, String nickname, String role) {
+        this.email = email;
+        this.password = password;
+        this.nickname = nickname;
+        this.role = UserRole.of(role);
+    }
+}
