+++ src/main/java/com/example/jwt/domain/user/repository/UserRepository.java
@@ -1,0 +1,16 @@
+package com.example.jwt.domain.user.repository;
+
+import com.example.jwt.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findUserByEmail(String email);
+
+    List<User> findUsersByIsDeleted(boolean isDeleted);
+}
