--- src/main/java/com/example/jwt/dto/LoginReqDto.java
@@ -1,19 +1,0 @@
-package com.example.jwt.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Getter;
-
-@Getter
-public class LoginReqDto {
-
-    @NotBlank(message = "email은 필수 항목 입니다.")
-    @Email(message = "이메일 형식이 아닙니다.")
-    private String email;
-
-    @NotBlank(message = "password는 필수 항목 입니다.")
-    @Pattern(regexp = "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$",
-            message = "비밀번호는 최소 8자 이상, 대문자, 소문자, 숫자, 특수문자를 포함해야 합니다.")
-    private String password;
-}
