--- src/main/java/com/example/jwt/dto/LoginResDto.java
@@ -1,13 +1,0 @@
-package com.example.jwt.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class LoginResDto {
-
-    private final String tokenAuthScheme;
-    private final String accessToken;
-    private final String refreshToken;
-}
