--- src/main/java/com/example/jwt/dto/UserDetailResDto.java
@@ -1,13 +1,0 @@
-package com.example.jwt.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class UserDetailResDto {
-
-    private final Long id;
-    private final String email;
-    private final String nickname;
-}
