--- src/main/java/com/example/jwt/dto/UserResDto.java
@@ -1,18 +1,0 @@
-package com.example.jwt.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@RequiredArgsConstructor
-public class UserResDto {
-
-    private final Long id;
-    private final String email;
-    private final String nickname;
-    private final String role;
-    private final LocalDateTime createdAt;
-    private final LocalDateTime modifiedAt;
-}
