--- src/main/java/com/example/jwt/error/CustomException.java
@@ -1,11 +1,0 @@
-package com.example.jwt.error;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class CustomException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-}
