--- src/main/java/com/example/jwt/error/ErrorResponse.java
@@ -1,30 +1,0 @@
-package com.example.jwt.error;
-
-import lombok.Getter;
-import org.springframework.http.ResponseEntity;
-
-import java.time.LocalDateTime;
-
-@Getter
-public class ErrorResponse {
-
-    private final LocalDateTime timestamp = LocalDateTime.now();
-    private final int status;
-    private final String error;
-    private final String code;
-    private final String message;
-
-    private ErrorResponse(ErrorCode errorCode) {
-        this.status = errorCode.getHttpStatus().value();
-        this.error = errorCode.getHttpStatus().getReasonPhrase();
-        this.code = errorCode.name();
-        this.message = errorCode.getMessage();
-    }
-
-    public static ResponseEntity<ErrorResponse> toResponseEntity(ErrorCode errorCode) {
-        ErrorResponse errorResponse = new ErrorResponse(errorCode);
-        return ResponseEntity
-                .status(errorCode.getHttpStatus())
-                .body(errorResponse);
-    }
-}
