+++ src/main/java/com/example/jwt/global/auth/filter/JwtAuthFilter.java
@@ -1,0 +1,88 @@
+package com.example.jwt.global.auth.filter;
+
+import com.example.jwt.global.auth.util.AuthenticationScheme;
+import com.example.jwt.global.auth.util.JwtProvider;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpHeaders;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+@Component
+@RequiredArgsConstructor
+public class JwtAuthFilter extends OncePerRequestFilter {
+
+    private final JwtProvider jwtProvider;
+    private final UserDetailsService userDetailsService;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request,
+                                    HttpServletResponse response,
+                                    FilterChain filterChain) throws ServletException, IOException {
+        String bearerToken = request.getHeader(HttpHeaders.AUTHORIZATION);
+        String headerPrefix = AuthenticationScheme.generateType(AuthenticationScheme.BEARER);
+
+        if (bearerToken == null || !bearerToken.startsWith(headerPrefix)) {
+            filterChain.doFilter(request, response);
+
+            return;
+        }
+
+        authenticate(request);
+        filterChain.doFilter(request, response);
+    }
+
+    /**
+     * request를 이용해 인증을 처리
+     * @param request
+     */
+    private void authenticate(HttpServletRequest request) {
+        String token = getTokenFromRequest(request);
+
+        if (!jwtProvider.validToken(token) || !jwtProvider.isAccessToken(token)) {
+            return;
+        }
+
+        String username = jwtProvider.getUsername(token);
+        UserDetails userDetails = userDetailsService.loadUserByUsername(username);
+
+        setAuthentication(request, userDetails);
+    }
+
+    /**
+     * request의 Authorization 헤더에서 Token 값을 추출
+     * @param request
+     * @return Token 값 (찾지 못하면 null)
+     */
+    private String getTokenFromRequest(HttpServletRequest request) {
+        final String bearerToken = request.getHeader(HttpHeaders.AUTHORIZATION);
+        final String headerPrefix = AuthenticationScheme.generateType(AuthenticationScheme.BEARER);
+
+        return StringUtils.hasText(bearerToken) && bearerToken.startsWith(headerPrefix)
+                ? bearerToken.substring(headerPrefix.length())
+                : null;
+    }
+
+    /**
+     * SecurityContext에 인증 객체를 저장
+     * @param request
+     * @param userDetails
+     */
+    private void setAuthentication(HttpServletRequest request, UserDetails userDetails) {
+        UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(userDetails, userDetails.getPassword(), userDetails.getAuthorities());
+        authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+}
