+++ src/main/java/com/example/jwt/global/error/CustomException.java
@@ -1,0 +1,11 @@
+package com.example.jwt.global.error;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class CustomException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+}
