+++ src/main/java/com/example/jwt/global/error/ErrorResponse.java
@@ -1,0 +1,30 @@
+package com.example.jwt.global.error;
+
+import lombok.Getter;
+import org.springframework.http.ResponseEntity;
+
+import java.time.LocalDateTime;
+
+@Getter
+public class ErrorResponse {
+
+    private final LocalDateTime timestamp = LocalDateTime.now();
+    private final int status;
+    private final String error;
+    private final String code;
+    private final String message;
+
+    private ErrorResponse(ErrorCode errorCode) {
+        this.status = errorCode.getHttpStatus().value();
+        this.error = errorCode.getHttpStatus().getReasonPhrase();
+        this.code = errorCode.name();
+        this.message = errorCode.getMessage();
+    }
+
+    public static ResponseEntity<ErrorResponse> toResponseEntity(ErrorCode errorCode) {
+        ErrorResponse errorResponse = new ErrorResponse(errorCode);
+        return ResponseEntity
+                .status(errorCode.getHttpStatus())
+                .body(errorResponse);
+    }
+}
