--- src/main/java/com/example/jwt/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.example.jwt.repository;
-
-import com.example.jwt.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findUserByEmail(String email);
-
-    List<User> findUsersByIsDeleted(boolean isDeleted);
-}
